corregido el problema de la paginacion
ahora las listas se muestran correctamente, falta implementar local storage correctamente

+++ src/storage.js
@@ -2,34 +2,36 @@
 export function savePokemonListOnStorage(lists) {
   lists.forEach((list) => {
     const { name } = list;
-    const { url } = list;
     try {
-      localStorage.setItem(name, JSON.stringify(url));
+      localStorage.setItem(name, JSON.stringify(list));
     } catch (e) {
       localStorage.clear();
-      localStorage.setItem(name, JSON.stringify(url));
+      localStorage.setItem(name, JSON.stringify(list));
     }
   });
   console.log(lists);
 }
 
-export function getPokemonListFromStorage(lists) {
-  const pokemonList = JSON.parse(localStorage.getItem(lists));
+export function getPokemonListFromStorage(name) {
+  const pokemonList = JSON.parse(localStorage.getItem(name));
   if (pokemonList === null) {
     throw new Error('PokemonList not found in Storage');
   }
-  console.log(lists);
+  console.log(name);
   return pokemonList;
 }
 
 export function savePokemonInfoOnStorage(pokemon) {
+  if (pokemon === undefined) {
+    throw new Error('Pokemon is undefined');
+  }
   try {
     localStorage.setItem(`${pokemon.name}`, JSON.stringify(pokemon));
-  } catch (e) {
+  } catch (error) {
     localStorage.clear();
     localStorage.setItem(`${pokemon.name}`, JSON.stringify(pokemon));
+    console.log(pokemon);
   }
-  console.log(pokemon);
 }
 
 export function getPokemonInfoFromStorage(name) {
